refactor: Product Page 위치 변경
